tests/wpt/tests/css/css-grid/grid-lanes/tentative/items/column-fit-content-percentage.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Masonry Layout Test: indefinite percentage in fit-content()</title>
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<link rel="help" href="https://drafts.csswg.org/css-grid/#track-sizes" title="7.2.1. Track Sizes">
<meta name="assert" content="Checks that an indefinite percentage in fit-content lets the masonry container grow enough to contain the max-content contribution of its masonry items.">
<style>
.container {
  width: 200px;
  margin-top: 10px;
}
.grid-lanes {
  display: inline-grid-lanes;
  background: blue;
}
.item {
  background: orange;
}
.item::before, .item::after {
  content: '';
  float: left;
  width: 50px;
  height: 50px;
}
</style>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>
"use strict";
function clamp(value, min, max) {
  return Math.max(min, Math.min(max, value));
}
const minContent = 50;
const maxContent = 100;
for (const use_min_width of [false, true]) {
  for (const use_calc of [false, true]) {
    for (const percentage of [0, 50, 75, 100, 150]) {
      const value = use_calc ? `fit-content(calc(0px + ${percentage}%))`
                            : `fit-content(${percentage}%)`;
      const container = document.createElement("div");
      container.className = "container";
      document.body.appendChild(container);
      const masonry = document.createElement("div");
      masonry.className = "grid-lanes";
      masonry.style.gridTemplateColumns = value;
      container.appendChild(masonry);
      const item = document.createElement("div");
      item.className = "item";
      item.style.minWidth = use_min_width ? "auto" : "0px";
      masonry.appendChild(item);
      test(function() {
        const minWidth = use_min_width ? minContent : 0;
        const colSize = clamp(percentage * maxContent / 100, minWidth, maxContent);
        const height = colSize < maxContent ? maxContent : minContent;
        assert_equals(item.offsetWidth, colSize, "Masonry item width");
        assert_equals(item.offsetHeight, height, "Masonry item height");
        assert_equals(masonry.offsetWidth, maxContent, "Masonry container width");
        assert_equals(masonry.offsetHeight, height, "Masonry container height");
        assert_equals(getComputedStyle(masonry).gridTemplateColumns, colSize + "px",
                      "Masonry column size");
      }, value + "; min-width: " + item.style.minWidth);
    }
  }
}
</script>
